refactor(server): split a file
